backend/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-autoslug==1.9.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
ecdsa==0.19.1
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
